Refactored into arrays using malloc, done this in order to hardcode the number of channels and there memory locations at run time, so all functions use the NUM_CHANNELS variable

+++ Firmware/Channels.h
@@ -13,7 +13,7 @@
 
 // Library Includes
 #include <Arduino.h>
-
+#include <stdint.h>
 /**
  * @brief Channel Selection Enum
  *
@@ -26,7 +26,7 @@ enum CHANNEL
   CHANNEL_3 = 3
 
 };
-#define NUM_CHANNELS 3
+#define NUM_CHANNELS (uint8_t)3
 
 /**
  * @brief Veroboard Pin Selections
